Projoint Example Data Set 1: "labelled tibble"

An already cleaned data frame (tibble) with each attribute corresponds to each column with the proper attribute name. The unit of observation must be The unit of observation must be each of two profiles in each task for each respondent.

Usage

exampleData1_labelled_tibble

Format

## `exampleData1_labelled_tibble` A data frame with 6,400 rows and 14 columns:

id

Unique respondent ID

task

Task number

profile

Profile number = {1,2}

selected

Whether or not a profile was selected

selected_repeated

Whether or not a profile was selected in the repeated task

School Quality, ... Type of Place

attributes

race

Respondent race: 6 categories

ideology

Respondent ideology from Extremely liberal to Extremely conservative

Source

Qualtrics and Prolific; see Clayton et al. replication materials.