Estimate and correct MMs or AMCEs

This is the internal function used to calculate and correct marginal means or average marginal component effects of a conjoint design.

Usage

pj_estimate(
  .data,
  .structure,
  .estimand,
  .att_choose,
  .lev_choose,
  .att_notchoose,
  .lev_notchoose,
  .att_choose_b,
  .lev_choose_b,
  .att_notchoose_b,
  .lev_notchoose_b,
  .se_method,
  .irr,
  .remove_ties,
  .ignore_position,
  .n_sims,
  .n_boot,
  .weights_1,
  .clusters_1,
  .se_type_1,
  .weights_2,
  .clusters_2,
  .se_type_2
)

Arguments

.data

A projoint_data object

.structure

Either "profile_level" (default) or "choice_level"

.estimand

Either "mm" for marginal mean or "amce" for average marginal component effect

.att_choose

A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels chosen).

.lev_choose

A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels chosen). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis

.att_notchoose

A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels not chosen). This argument should be specified only if the .structure argument is "choice-level".

.lev_notchoose

A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels not chosen). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis. This argument should be specified only if the .structure argument is "choice-level".

.att_choose_b

[baseline for AMCE] A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels chosen).

.lev_choose_b

[baseline for AMCE] A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels chosen). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis

.att_notchoose_b

[baseline for AMCE] A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels not chosen*). This argument should be specified only if the .structure argument is "choice-level".

.lev_notchoose_b

[baseline for AMCE] A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels not chosen). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis. This argument should be specified only if the .structure argument is "choice-level".

.se_method

By default, c("analytic", "simulation", "bootstrap") description

.irr

NULL (default) if IRR is to be calculated using the repeated task. Otherwise, a numerical value

.remove_ties

Logical: should ties be removed before estimation? Defaults to TRUE.

.ignore_position

TRUE (default) if you ignore the location of profile (left or right). Relevant only if analyzed at the choice level

.n_sims

The number of simulations. Relevant only if .se_method == "simulation"

.n_boot

The number of bootstrapped samples. Relevant only if .se_method == "bootstrap"

.weights_1

the weight to estimate IRR (see lm_robust): NULL (default)

.clusters_1

the clusters to estimate IRR (see lm_robust): NULL (default)

.se_type_1

the standard error type to estimate IRR (see lm_robust): "classical" (default)

.weights_2

the weight to estimate MM or AMCE (see lm_robust): NULL (default)

.clusters_2

the clusters to estimate MM or AMCE (see lm_robust): NULL (default)

.se_type_2

the standard error type to estimate MM or AMCE (see lm_robust): "classical" (default)

Value

A data frame of estimates