Analyze a conjoint data set and correct for measurement error

This main function analyzes a conjoint data set and produces measurement error-corrected estimates of either marginal means or average marginal component effects, ready for plotting. It accepts a `projoint_data` object, and optionally a `projoint_qoi` object for users who wish to specify more complex quantities of interest.

Usage

projoint(
  .data,
  .qoi = NULL,
  .structure = "profile_level",
  .estimand = "mm",
  .se_method = "analytical",
  .irr = NULL,
  .remove_ties = TRUE,
  .ignore_position = NULL,
  .n_sims = NULL,
  .n_boot = NULL
)

Arguments

.data

A `projoint_data` object

.qoi

A `projoint_qoi` object. If NULL, defaults to producing all MMs and all AMCEs.

.structure

Either "profile_level" or "choice_level"

.estimand

Either "mm" for marginal mean or "amce" for average marginal component effect

.se_method

c("analytic", "simulation", "bootstrap") description

.irr

NULL (default) if IRR is to be calculated using the repeated task. Otherwise, a numerical value

.remove_ties

Logical: should ties be removed before estimation? Defaults to TRUE.

.ignore_position

TRUE (default) if you ignore the location of profile (left or right. Relevant only if analyzed at the choice level

.n_sims

The number of simulations. Relevant only if .se_method == "simulation"

.n_boot

The number of bootstrapped samples. Relevant only if .se_method == "bootstrap"

Value

A `projoint_results` object

Examples

library(projoint)

data("exampleData1")
head(exampleData1)
#> # A tibble: 6 × 185
#>   ResponseId      choice1_repeated_fli…¹ choice1 choice2 choice3 choice4 choice5
#>   <chr>           <chr>                  <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 R_1M3TDihZzq9z… Community B            Commun… Commun… Commun… Commun… Commun…
#> 2 R_3HtXzkcSSlfi… Community B            Commun… Commun… Commun… Commun… Commun…
#> 3 R_yjYj0jtOY98X… Community B            Commun… Commun… Commun… Commun… Commun…
#> 4 R_1dKd05O6FTOV… Community B            Commun… Commun… Commun… Commun… Commun…
#> 5 R_1otDp642wWYl… Community A            Commun… Commun… Commun… Commun… Commun…
#> 6 R_2BnD3fuJMRKZ… Community A            Commun… Commun… Commun… Commun… Commun…
#> # ℹ abbreviated name: ¹​choice1_repeated_flipped
#> # ℹ 178 more variables: choice6 <chr>, choice7 <chr>, choice8 <chr>,
#> #   race <chr>, party_1 <chr>, party_2 <chr>, party_3 <chr>, party_4 <chr>,
#> #   ideology <chr>, honesty <chr>, `K-1-1` <chr>, `K-1-1-1` <chr>,
#> #   `K-1-2` <chr>, `K-1-1-2` <chr>, `K-1-3` <chr>, `K-1-1-3` <chr>,
#> #   `K-1-4` <chr>, `K-1-1-4` <chr>, `K-1-5` <chr>, `K-1-1-5` <chr>,
#> #   `K-1-6` <chr>, `K-1-1-6` <chr>, `K-1-7` <chr>, `K-1-1-7` <chr>, …

outcomes <- paste0("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")

reshaped_data <- reshape_projoint(
  .dataframe = exampleData1, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = TRUE)

projoint(reshaped_data) 
#> An object of class "projoint_results_mm"
#> Slot "estimates":
#> # A tibble: 48 × 5
#>    estimand       estimate     se attribute level 
#>    <chr>             <dbl>  <dbl> <chr>     <chr> 
#>  1 mm_uncorrected    0.574 0.0134 att1      level1
#>  2 mm_corrected      0.614 0.0207 att1      level1
#>  3 mm_uncorrected    0.485 0.0134 att1      level2
#>  4 mm_corrected      0.477 0.0204 att1      level2
#>  5 mm_uncorrected    0.445 0.0131 att1      level3
#>  6 mm_corrected      0.416 0.0203 att1      level3
#>  7 mm_uncorrected    0.489 0.0133 att2      level1
#>  8 mm_corrected      0.483 0.0202 att2      level1
#>  9 mm_uncorrected    0.524 0.0130 att2      level2
#> 10 mm_corrected      0.536 0.0200 att2      level2
#> # ℹ 38 more rows
#> 
#> Slot "labels":
#> # A tibble: 24 × 4
#>    attribute                level                        attribute_id level_id  
#>    <chr>                    <chr>                        <chr>        <chr>     
#>  1 Housing Cost             15% of pre-tax income        att1         att1:leve…
#>  2 Housing Cost             30% of pre-tax income        att1         att1:leve…
#>  3 Housing Cost             40% of pre-tax income        att1         att1:leve…
#>  4 Presidential Vote (2020) 30% Democrat, 70% Republican att2         att2:leve…
#>  5 Presidential Vote (2020) 50% Democrat, 50% Republican att2         att2:leve…
#>  6 Presidential Vote (2020) 70% Democrat, 30% Republican att2         att2:leve…
#>  7 Racial Composition       50% White, 50% Nonwhite      att3         att3:leve…
#>  8 Racial Composition       75% White, 25% Nonwhite      att3         att3:leve…
#>  9 Racial Composition       90% White, 10% Nonwhite      att3         att3:leve…
#> 10 Racial Composition       96% White, 4% Nonwhite       att3         att3:leve…
#> # ℹ 14 more rows
#> 
#> Slot "data":
#> # A tibble: 6,400 × 13
#>    id            task profile att4  att7  att3  att1  att2  att5  att6  selected
#>    <chr>        <dbl>   <dbl> <fct> <fct> <fct> <fct> <fct> <fct> <fct>    <dbl>
#>  1 R_1M3TDihZz…     1       1 att4… att7… att3… att1… att2… att5… att6…        0
#>  2 R_1M3TDihZz…     1       2 att4… att7… att3… att1… att2… att5… att6…        1
#>  3 R_1M3TDihZz…     2       1 att4… att7… att3… att1… att2… att5… att6…        0
#>  4 R_1M3TDihZz…     2       2 att4… att7… att3… att1… att2… att5… att6…        1
#>  5 R_1M3TDihZz…     3       1 att4… att7… att3… att1… att2… att5… att6…        0
#>  6 R_1M3TDihZz…     3       2 att4… att7… att3… att1… att2… att5… att6…        1
#>  7 R_1M3TDihZz…     4       1 att4… att7… att3… att1… att2… att5… att6…        1
#>  8 R_1M3TDihZz…     4       2 att4… att7… att3… att1… att2… att5… att6…        0
#>  9 R_1M3TDihZz…     5       1 att4… att7… att3… att1… att2… att5… att6…        0
#> 10 R_1M3TDihZz…     5       2 att4… att7… att3… att1… att2… att5… att6…        1
#> # ℹ 6,390 more rows
#> # ℹ 2 more variables: selected_repeated <dbl>, agree <dbl>
#> 
#> Slot "attribute_of_interest":
#> NULL
#> 
#> Slot "levels_of_interest":
#> NULL
#> 
#> Slot "irr":
#> [1] 0.1721281
#> 
#> Slot "figure":
#> NULL
#>