Analyze a conjoint data set and correct for measurement error

This function ...

Usage

projoint(
  .data,
  .qoi = NULL,
  .structure = "profile_level",
  .estimand = "mm",
  .se_method = "analytical",
  .irr = NULL,
  .remove_ties = TRUE,
  .ignore_position = NULL,
  .n_sims = NULL,
  .n_boot = NULL
)

Arguments

.data

A `projoint_data` object

.qoi

A `projoint_qoi` object. If NULL, defaults to producing all MMs and all AMCEs.

.structure

Either "profile_level" or "choice_level"

.estimand

Either "mm" for marginal mean or "amce" for average marginal component effect

.se_method

c("analytic", "simulation", "bootstrap") description

.irr

NULL (default) if IRR is to be calculated using the repeated task. Otherwise, a numerical value

.remove_ties

Logical: should ties be removed before estimation? Defaults to TRUE.

.ignore_position

TRUE (default) if you ignore the location of profile (left or right. Relevant only if analyzed at the choice level

.n_sims

The number of simulations. Relevant only if .se_method == "simulation"

.n_boot

The number of bootstrapped samples. Relevant only if .se_method == "bootstrap"

Value

A `projoint_results` object