Set the quantities of interest

This function allows users to specify more fine-tuned details of their conjoint analysis. In particular, users can set very specific quantities of interest aside from simple AMCEs or MMs.

Usage

set_qoi(
  .structure = "choice_level",
  .estimand = "mm",
  .att_choose,
  .lev_choose,
  .att_notchoose = NULL,
  .lev_notchoose = NULL,
  .att_choose_b = NULL,
  .lev_choose_b = NULL,
  .att_notchoose_b = NULL,
  .lev_notchoose_b = NULL
)

Arguments

.structure

Either "choice_level" (default) or "profile_level"

.estimand

Either "mm" for marginal mean or "amce" for average marginal component effect

.att_choose

A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels chosen).

.lev_choose

A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels chosen). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis

.att_notchoose

A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels not chosen). This argument should be specified only if the .structure argument is "choice-level".

.lev_notchoose

A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels not chosen). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis. This argument should be specified only if the .structure argument is "choice-level".

.att_choose_b

[baseline for AMCE] A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels chosen).

.lev_choose_b

[baseline for AMCE] A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels chosen). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis

.att_notchoose_b

[baseline for AMCE] A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels not chosen*). This argument should be specified only if the .structure argument is "choice-level".

.lev_notchoose_b

[baseline for AMCE] A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels not chosen). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis. This argument should be specified only if the .structure argument is "choice-level".

Value

A projoint_qoi object